Use a SubUpdateResult to pass messages back to main
When a sub update function needs to resend a message instead of command,
return a Message msg instead of a command. The outer update function
will then recursively call update with the new message.

This is not perfect as those messages aren't logged by the debugger and
there might be other issues as well.